Added sim2sim helpers.py
